requirements.txt
Django==4.2.7
django-unfold==0.18.0
Pillow==10.1.0